Fix compilation issues on Mac OS X
The upgrade to XCODE 5.1 introduced an extra bug where the clang flag
-multiply_definedsuppress was removed and will no longer be supported.
Due to not being able to compile on my linux computer at the moment I
have included the change to DLDFLAGS in the logic to detect if the user
is compiling on a Mac OS. The following is the error generated without
the DLDFLAGS:

```
clang: error: unknown argument: '-multiply_definedsuppress'
[-Wunused-command-line-argument-hard-error-in-future]
clang: note: this will be a hard error (cannot be downgraded to a
warning) in the future
```

The other bugs this fixes is the inability to compile since Mac OS X does
not have a /usr/pkg/lib directory.

That is all.